Refactoring out operations for clarity
 - Renamed operation to have 6502 prefix
 - Renamed addressing mode to have 6502 prefix
