+++ src/main/java/com/rox/emu/processor/mos6502/op/Mos6502Instruction.java
@@ -5,7 +5,7 @@ import com.rox.emu.processor.mos6502.Mos6502Alu;
 import com.rox.emu.processor.mos6502.Registers;
 
 @FunctionalInterface
-public interface Instruction {
+public interface Mos6502Instruction {
     void perform(final Mos6502Alu alu,
                  final Registers registers,
                  final Memory memory);
