+++ src/main/java/com/rox/emu/processor/mos6502/op/util/OpCodeConverter.java
@@ -2,12 +2,11 @@ package com.rox.emu.processor.mos6502.op.util;
 
 import com.rox.emu.UnknownOpCodeException;
 import com.rox.emu.processor.mos6502.Mos6502;
-import com.rox.emu.processor.mos6502.op.AddressingMode;
+import com.rox.emu.processor.mos6502.op.Mos6502AddressingMode;
 import com.rox.emu.processor.mos6502.op.OpCode;
+import com.rox.emu.processor.mos6502.op.Mos6502Operation;
 
 import java.util.EnumSet;
-
-import static com.rox.emu.processor.mos6502.op.OpCode.*;
 
 
 /**
@@ -34,26 +33,26 @@ public class OpCodeConverter {
 
     /**
      * @param internalOpCodeName {@link String} representation of an {@link OpCode}
-     * @return the {@link Operation} associated with this {@link OpCode}
+     * @return the {@link Mos6502Operation} associated with this {@link OpCode}
      */
-    public static Operation getOperation(final String internalOpCodeName) {
-        return Operation.valueOf(getOpCode(internalOpCodeName));
+    public static Mos6502Operation getOperation(final String internalOpCodeName) {
+        return Mos6502Operation.valueOf(getOpCode(internalOpCodeName));
     }
 
     /**
-     * Extract the {@link Mos6502} {@link AddressingMode} from the {@link String} representation of an {@link OpCode}.
+     * Extract the {@link Mos6502} {@link Mos6502AddressingMode} from the {@link String} representation of an {@link OpCode}.
      *
      * @param internalOpCodeName the {@link String} representation of an {@link OpCode}
-     * @return An {@link AddressingMode} object that represents the intended addressing mode of the {@link OpCode} in question
+     * @return An {@link Mos6502AddressingMode} object that represents the intended addressing mode of the {@link OpCode} in question
      */
-    public static AddressingMode getAddressingMode(String internalOpCodeName){
+    public static Mos6502AddressingMode getAddressingMode(String internalOpCodeName){
         final String[] tokens = internalOpCodeName.split(OpCode.TOKEN_SEPARATOR);
         if (tokens.length <= OpCode.ADDR_I) {
             //XXX Write this less ugly
-            if (EnumSet.of(Operation.JSR, Operation.BPL, Operation.BMI, Operation.BVC, Operation.BVS, Operation.BCC, Operation.BCS, Operation.BNE, Operation.BEQ).contains(getOperation(internalOpCodeName)))
-                return AddressingMode.RELATIVE;
+            if (EnumSet.of(Mos6502Operation.JSR, Mos6502Operation.BPL, Mos6502Operation.BMI, Mos6502Operation.BVC, Mos6502Operation.BVS, Mos6502Operation.BCC, Mos6502Operation.BCS, Mos6502Operation.BNE, Mos6502Operation.BEQ).contains(getOperation(internalOpCodeName)))
+                return Mos6502AddressingMode.RELATIVE;
 
-            return AddressingMode.IMPLIED;
+            return Mos6502AddressingMode.IMPLIED;
         }
 
         final String addressingModeDescriptor = tokens[OpCode.ADDR_I];
@@ -61,20 +60,20 @@ public class OpCodeConverter {
         //XXX Not pretty but necessary for proper test coverage - update if JaCoCo ever learns to deal with it
         final String indexToken = (tokens.length <= OpCode.INDX_I) ? "" : tokens[OpCode.INDX_I];
         if ("I".equals(addressingModeDescriptor))
-            return AddressingMode.IMMEDIATE;
+            return Mos6502AddressingMode.IMMEDIATE;
         else if ("A".equals(addressingModeDescriptor))
-            return AddressingMode.ACCUMULATOR;
+            return Mos6502AddressingMode.ACCUMULATOR;
         else if ("Z".equals(addressingModeDescriptor))
-            return withIndexing(AddressingMode.ZERO_PAGE, indexToken);
+            return withIndexing(Mos6502AddressingMode.ZERO_PAGE, indexToken);
         else if ("ABS".equals(addressingModeDescriptor))
-            return withIndexing(AddressingMode.ABSOLUTE, indexToken);
+            return withIndexing(Mos6502AddressingMode.ABSOLUTE, indexToken);
         else if ("IND".equals(addressingModeDescriptor))
-            return withIndexing(AddressingMode.INDIRECT, indexToken);
+            return withIndexing(Mos6502AddressingMode.INDIRECT, indexToken);
         else
             throw new UnknownOpCodeException("Unrecognised addressing mode " + addressingModeDescriptor, internalOpCodeName);
     }
 
-    private static AddressingMode withIndexing(final AddressingMode addressingMode, final String indexToken){
+    private static Mos6502AddressingMode withIndexing(final Mos6502AddressingMode addressingMode, final String indexToken){
         if ("IX".equalsIgnoreCase(indexToken))
             return addressingMode.xIndexed();
         else if ("IY".equalsIgnoreCase(indexToken))
