+++ src/test/java/com/rox/emu/processor/mos6502/op/util/AddressingModeTest.java
@@ -1,7 +1,7 @@
 package com.rox.emu.processor.mos6502.op.util;
 
 import com.rox.emu.UnknownOpCodeException;
-import com.rox.emu.processor.mos6502.op.AddressingMode;
+import com.rox.emu.processor.mos6502.op.Mos6502AddressingMode;
 import org.junit.Test;
 
 import static junit.framework.TestCase.fail;
@@ -10,7 +10,7 @@ import static org.junit.Assert.*;
 public class AddressingModeTest {
     @Test
     public void testValues(){
-        for (AddressingMode addressingMode : AddressingMode.values()){
+        for (Mos6502AddressingMode addressingMode : Mos6502AddressingMode.values()){
             final String description = addressingMode.getDescription();
             final int bytes = addressingMode.getInstructionBytes();
 
@@ -21,15 +21,15 @@ public class AddressingModeTest {
 
     @Test
     public void testConversionToXIndex(){
-        assertEquals(AddressingMode.ZERO_PAGE_X, AddressingMode.ZERO_PAGE.xIndexed());
-        assertEquals(AddressingMode.ABSOLUTE_X, AddressingMode.ABSOLUTE.xIndexed());
-        assertEquals(AddressingMode.INDIRECT_X, AddressingMode.INDIRECT.xIndexed());
+        assertEquals(Mos6502AddressingMode.ZERO_PAGE_X, Mos6502AddressingMode.ZERO_PAGE.xIndexed());
+        assertEquals(Mos6502AddressingMode.ABSOLUTE_X, Mos6502AddressingMode.ABSOLUTE.xIndexed());
+        assertEquals(Mos6502AddressingMode.INDIRECT_X, Mos6502AddressingMode.INDIRECT.xIndexed());
     }
 
     @Test
     public void testInvalidConversionToXIndex(){
         try {
-            AddressingMode.ACCUMULATOR.xIndexed();
+            Mos6502AddressingMode.ACCUMULATOR.xIndexed();
             fail("Accumulator cannot be X indexed");
         }catch(UnknownOpCodeException e){
             assertNotNull(e);
@@ -39,15 +39,15 @@ public class AddressingModeTest {
 
     @Test
     public void testConversionToYIndex(){
-        assertEquals(AddressingMode.ZERO_PAGE_Y, AddressingMode.ZERO_PAGE.yIndexed());
-        assertEquals(AddressingMode.ABSOLUTE_Y, AddressingMode.ABSOLUTE.yIndexed());
-        assertEquals(AddressingMode.INDIRECT_Y, AddressingMode.INDIRECT.yIndexed());
+        assertEquals(Mos6502AddressingMode.ZERO_PAGE_Y, Mos6502AddressingMode.ZERO_PAGE.yIndexed());
+        assertEquals(Mos6502AddressingMode.ABSOLUTE_Y, Mos6502AddressingMode.ABSOLUTE.yIndexed());
+        assertEquals(Mos6502AddressingMode.INDIRECT_Y, Mos6502AddressingMode.INDIRECT.yIndexed());
     }
 
     @Test
     public void testInvalidConversionToYIndex(){
         try {
-            AddressingMode.ACCUMULATOR.yIndexed();
+            Mos6502AddressingMode.ACCUMULATOR.yIndexed();
             fail("Accumulator cannot be Y indexed");
         }catch(UnknownOpCodeException e){
             assertNotNull(e);
@@ -57,7 +57,7 @@ public class AddressingModeTest {
 
     @Test
     public void testToStringIsACompleteDescription(){
-        for (AddressingMode addressingMode : AddressingMode.values()) {
+        for (Mos6502AddressingMode addressingMode : Mos6502AddressingMode.values()) {
             for (String descriptiveWord : addressingMode.name().split("_")) {
                 assertTrue(addressingMode.toString().toLowerCase().contains(descriptiveWord.toLowerCase()));
             }
